requirements.txt
streamlit
requests
python-docx
pymupdf
deepgram-sdk
google-generativeai
python-dotenv
